requirements.txt
joblib==1.4.2
matplotlib==3.7.2
numpy==2.3.5
pandas==2.3.3
plotly==6.3.1
Pyrebase4==4.8.0
python-dotenv==1.2.1
scikit_learn==1.7.2
seaborn==0.13.2
streamlit==1.43.2
xgboost==3.1.2


